minor tweaks to yelp query form based on merged changes
